+++ lib/src/quds_db_mysql_base.dart
@@ -2,3 +2,6 @@ export 'adapters/mysql_database_adapter.dart';
 export 'adapters/mysql_database_connection.dart';
 export 'providers/mysql_table_provider.dart';
 export 'builders/mysql_query_builder.dart';
+export 'schema/mysql_schema_inspector.dart';
+export 'schema/mysql_schema_migrator.dart';
+export 'migration/mysql_migration_context.dart';